requirements.txt
## The following requirements were added by pip freeze:
grpcio-tools==1.59.2
protobuf==4.25.0
six==1.16.0
